Refactor MonitorRunner (#363)
* Add InputService

* Change MonitorRunner to singleton and use register methods for components

* Update MonitorRunner to use InputService

* Add TriggerService

* Update MonitorRunner to use TriggerService

* Add AlertService

* Update MonitorRunner to use AlertService

* Update comment in AlertService
