+++ alerting/src/main/kotlin/com/amazon/opendistroforelasticsearch/alerting/settings/DestinationSettings.kt
@@ -34,6 +34,7 @@ class DestinationSettings {
         const val EMAIL_DESTINATION_SETTING_PREFIX = DESTINATION_SETTING_PREFIX + "email."
         val ALLOW_LIST_ALL = DestinationType.values().toList().map { it.value }
         val ALLOW_LIST_NONE = emptyList<String>()
+        val HOST_DENY_LIST_NONE = emptyList<String>()
 
         val ALLOW_LIST: Setting<List<String>> = Setting.listSetting(
             DESTINATION_SETTING_PREFIX + "allow_list",
@@ -58,11 +59,11 @@ class DestinationSettings {
         )
 
         val HOST_DENY_LIST: Setting<List<String>> = Setting.listSetting(
-                "opendistro.destination.host.deny_list",
-                emptyList<String>(),
-                Function.identity(),
-                Setting.Property.NodeScope,
-                Setting.Property.Final
+            "opendistro.destination.host.deny_list",
+            HOST_DENY_LIST_NONE,
+            Function.identity(),
+            Setting.Property.NodeScope,
+            Setting.Property.Final
         )
 
         fun loadDestinationSettings(settings: Settings): Map<String, SecureDestinationSettings> {
